cs3240-s16-team13/requirements.txt
Django==1.8
dj-database-url==0.4.0
dj-static==0.0.6
django-bootstrap3==7.0.1
django-toolbelt==0.0.1
gunicorn==19.4.5
psycopg2==2.4.5
pycrypto==2.6.1
static3==0.6.1
whitenoise==3.0
Pillow==3.1.1